+++ examples/connectivity/plot_cwt_sensor_connectivity.py
@@ -41,7 +41,8 @@ events = mne.read_events(event_fname)
 raw.info['bads'] += ['MEG 2443']
 
 # Pick MEG gradiometers
-picks = fiff.pick_types(raw.info, meg='grad', eeg=False, stim=False, eog=True)
+picks = fiff.pick_types(raw.info, meg='grad', eeg=False, stim=False, eog=True,
+                        exclude='bads')
 
 # Create epochs for left-visual condition
 event_id, tmin, tmax = 3, -0.2, 0.5