+++ examples/export/plot_epochs_as_data_frame.py
@@ -110,7 +110,7 @@ events = mne.read_events(event_fname)[:10]
 # Add a bad channel
 raw.info['bads'] += ['MEG 2443']
 picks = mne.fiff.pick_types(raw.info, meg='grad', eeg=False, eog=True,
-                            stim=False)
+                            stim=False, exclude='bads')
 
 tmin, tmax = -0.2, 0.5
 baseline = (None, 0)