+++ examples/plot_simulate_evoked_data.py
@@ -68,7 +68,7 @@ stc = generate_sparse_stc(fwd['src'], labels, stc_data, tmin, tstep,
 
 ###############################################################################
 # Generate noisy evoked data
-picks = mne.fiff.pick_types(raw.info, meg=True)
+picks = mne.fiff.pick_types(raw.info, meg=True, exclude='bads')
 iir_filter = iir_filter_raw(raw, order=5, picks=picks, tmin=60, tmax=180)
 evoked = generate_evoked(fwd, stc, evoked_template, cov, snr,
                          tmin=0.0, tmax=0.2, iir_filter=iir_filter)