+++ examples/plot_topo_compare_conditions.py
@@ -48,7 +48,7 @@ reject = dict(grad=4000e-13, mag=4e-12)
 
 # pick MEG channels
 picks = pick_types(raw.info, meg=True, eeg=False, stim=False, eog=True,
-                   include=include)
+                   include=include, exclude='bads')
 
 # Create epochs including different events
 epochs = mne.Epochs(raw, events, dict(audio_l=1, visual_r=3), tmin, tmax,