+++ examples/stats/plot_cluster_1samp_test_time_frequency.py
@@ -47,7 +47,7 @@ raw.info['bads'] += ['MEG 2443', 'EEG 053']  # bads + 2 more
 
 # picks MEG gradiometers
 picks = fiff.pick_types(raw.info, meg='grad', eeg=False, eog=True,
-                                stim=False, include=include)
+                                stim=False, include=include, exclude='bads')
 
 # Load condition 1
 event_id = 1