+++ mne/epochs.py
@@ -235,7 +235,7 @@ class Epochs(object):
         self.info = cp.deepcopy(raw.info)
         if picks is None:
             picks = pick_types(raw.info, meg=True, eeg=True, stim=True,
-                        ecg=True, eog=True, misc=True)
+                               ecg=True, eog=True, misc=True, exclude='bads')
         self.info['chs'] = [self.info['chs'][k] for k in picks]
         self.info['ch_names'] = [self.info['ch_names'][k] for k in picks]
         self.ch_names = self.info['ch_names']
@@ -393,7 +393,7 @@ class Epochs(object):
         # Detrend
         if self.detrend is not None:
             picks = pick_types(self.info, meg=True, eeg=True, stim=False,
-                               eog=False, ecg=False, emg=False)
+                               eog=False, ecg=False, emg=False, exclude=[])
             epoch[picks] = detrend(epoch[picks], self.detrend, axis=1)
 
         # Baseline correct
@@ -672,7 +672,7 @@ class Epochs(object):
         if picks is None:
             picks = pick_types(evoked.info, meg=True, eeg=True,
                                stim=False, eog=False, ecg=False,
-                               emg=False)
+                               emg=False, exclude=[])
             if len(picks) == 0:
                 raise ValueError('No data channel found when averaging.')
 
@@ -969,7 +969,8 @@ class Epochs(object):
             raise Exception('the nitime package is missing')
 
         if picks is None:
-            picks = pick_types(self.info, include=self.ch_names)
+            picks = pick_types(self.info, include=self.ch_names,
+                               exclude='bads')
         if epochs_idx is None:
             epochs_idx = slice(len(self.events))